requirements.txt
yfinance==0.2.65
Flask==3.1.1
python-dotenv==1.0.1
twilio==9.7.0
schedule==1.2.1
pytz==2024.1
